feat(design-tracker): consolidate bulk actions into a single Bulk Update dialog
- Replaced separate `BulkAssignDialog` and `BulkStatusDialog` components with a unified `BulkUpdateDialog`.
- Refactored `project-design-tracker-details.tsx` to handle designers, deadlines, and status changes through a single `handleBulkUpdate` method.
- Updated `bulk_update_task_status.py` to correctly check for the Admin role using the user's `role_profile_name`.
- Added an `ignore_design_tracker_status_validation` flag to `design_tracker_task_child_table.py` to bypass evidence requirements (like `file_link` and `approval_proof`) during Admin bulk status overrides.
